+++ resourcet-extra/default.nix
@@ -3,7 +3,7 @@
 }:
 mkDerivation {
   pname = "resourcet-extra";
-  version = "0.0.2";
+  version = "0.1.0";
   src = ./.;
   libraryHaskellDepends = [
     async base containers resourcet safe-exceptions unliftio-core