+++ resourcet-extra/CHANGELOG.md
@@ -1,4 +1,6 @@
-# Version 0.0.2
+# Version 0.1.0
+
+* Fix `onceK`, `once`.
 
 * Add `runResourceT`, `asyncRestore`, `releaseType`, `unprotectType`.